lib_string.c: replace the loop counter and argument type to uint.

+++ src/lib_string.c
@@ -91,9 +91,9 @@ int strcmp_w(UTF16 a, UTF16 b)
 }
 
 __declspec(noinline)
-int strncmp_a(ANSI a, ANSI b, int64 n)
+int strncmp_a(ANSI a, ANSI b, uint n)
 {
-    for (int64 i = 0; i < n; i++)
+    for (uint i = 0; i < n; i++)
     {
         byte s0 = *a;
         byte s1 = *b;
@@ -118,9 +118,9 @@ int strncmp_a(ANSI a, ANSI b, int64 n)
 }
 
 __declspec(noinline)
-int strncmp_w(UTF16 a, UTF16 b, int64 n)
+int strncmp_w(UTF16 a, UTF16 b, uint n)
 {
-    for (int64 i = 0; i < n; i++)
+    for (uint i = 0; i < n; i++)
     {
         uint16 s0 = *a;
         uint16 s1 = *b;
@@ -197,9 +197,9 @@ int stricmp_w(UTF16 a, UTF16 b)
 }
 
 __declspec(noinline)
-int strnicmp_a(ANSI a, ANSI b, int64 n)
+int strnicmp_a(ANSI a, ANSI b, uint n)
 {
-    for (int64 i = 0; i < n; i++)
+    for (uint i = 0; i < n; i++)
     {
         byte s0 = lowercase_a(*a);
         byte s1 = lowercase_a(*b);
@@ -224,9 +224,9 @@ int strnicmp_a(ANSI a, ANSI b, int64 n)
 }
 
 __declspec(noinline)
-int strnicmp_w(UTF16 a, UTF16 b, int64 n)
+int strnicmp_w(UTF16 a, UTF16 b, uint n)
 {
-    for (int64 i = 0; i < n; i++)
+    for (uint i = 0; i < n; i++)
     {
         uint16 s0 = lowercase_w(*a);
         uint16 s1 = lowercase_w(*b);
@@ -311,10 +311,10 @@ uint strcpy_w(UTF16 dst, UTF16 src)
 }
 
 __declspec(noinline)
-uint strncpy_a(ANSI dst, ANSI src, int64 n)
+uint strncpy_a(ANSI dst, ANSI src, uint n)
 {
     uint l = 0;
-    for (int64 i = 0; i < n; i++)
+    for (uint i = 0; i < n; i++)
     {
         byte s = *src;
         *dst = s;
@@ -331,10 +331,10 @@ uint strncpy_a(ANSI dst, ANSI src, int64 n)
 }
 
 __declspec(noinline)
-uint strncpy_w(UTF16 dst, UTF16 src, int64 n)
+uint strncpy_w(UTF16 dst, UTF16 src, uint n)
 {
     uint l = 0;
-    for (int64 i = 0; i < n; i++)
+    for (uint i = 0; i < n; i++)
     {
         uint16 s = *src;
         *dst = s;
